Move Bootstrap.class to presto-bytecode to avoid duplicate definitions
